Performance optimizations (XML Archiving) and DataSource fixes
Increased performance using Xml Archiving.

Fixed DataSource Definition validation which caused "PHP Notice: Undefined index" to be emitted to the log.

Moved NextOrderId and NextInvoiceId to new SMShopState DataSource since SMAttributes locking is not guaranteed to result in consistent increments - other extensions are not forced to also call SMAttributes::Lock() so the lock set by SMShop might be ignored.

At what time an order confirmation e-mail is sent is now determined by whether a payment method was selected when placing the order. If no payment method is selected (when no payment methods are configured) the confirmation e-mail is delivered immediately. If a payment method on the other hand is selected, the confirmation e-mail is delivered when the payment is authorized.

Added "Working on it.." dialog to OrderForm which emerges 8 seconds after placing an order, in case the server is very slow. Also added an animated spinner icon to the proceed button to immediately indicate that the system is processing the order.

The Base Model in JSShop can now trigger async maintainance jobs using the headers Post-Process-Url, Post-Process-Payload, and Post-Process-Arguments.