Concurrency improvement for DataSource callback.
DataSource callback did not allow multiple concurrent requests - the only exception was the XML Archiving operation which was allowed to run simultaneously with other requests. But if the user triggered a series of unrelated operations, and two of those caused locking on a data source, all subsequent operations would be stalling due to locking on session state, making it impossible to even browse the website. A concrete example happend during testing where placing an order triggered XML Archiving. Browsing to the OrderList then caused a subsequent request to load order data which had to wait for the XML Archiving operation to finish due to DataSource locking, leaving a session open that prevented any other request from being processed, making it impossible to even load web pages. To prevent similar problems in the future, the DataSource callback is now stateless, meaning session state has been disabled/closed so that PHP does not cause locking.