Fixed XML Archiving which gradually unlocked DataSources.
XML Archiving gradually released locks on DataSources when data was committed which is not transaction safe. This bug has now been fixed so that locks are released after all DataSources have been committed, ensuring that other sessions do not read data while it's in an inconsistent state ("all or nothing").

Also fixed bug related to unlocking. Locks were only explicitly released if XML Archiving operation ran which in some cases left it to the OS to release the locks. Leaving unlocking to the OS may result in delayed unlocking, potentially blocking other sessions for a while.