Fixed error when updating existing row with non-existing field.
Adding a new column/field/attribute to a data source caused an error to be thrown when the new column/field/attribute was being added to an existing row using Update(..).
The internal function getCaseSensitiveAttribute(..) returned Null which caused an error to be thrown. The function therefore no longer returns Null if the row does not contain the specified column/field/attribute - instead it simply returns the name of the field in the casing it was provided in.